switch to using dockerfile plugin for builds
There is some magic in the docker-maven-plugin that obscures what is
actually in the image.  This switches to the dockerfile-maven-plugin
which utilizes a Dockerfile per app to build the resultant image.

This way you can be a lot more obvious about what is inside the image,
and it allows ops/security/etc a way to shape the underlying base image
to contain appropriate compliance items.